init repository layer (#26)
