+++ src/lib.rs
@@ -1,10 +1,12 @@
 pub mod domain;
+pub mod repositories;
 pub mod routes;
 pub mod startup;
 pub mod usecases;
 
 use shaku::module;
-use startup::MySqlDatabase;
+
+use repositories::MySqlDatabase;
 use usecases::{health_check::HealthCheckUseCase, signup::SignUpUseCase};
 
 module! {
