+++ src/startup.rs
@@ -1,16 +1,13 @@
-use std::str::FromStr;
 use std::sync::Arc;
 
 use actix_settings::{ApplySettings as _, Settings};
 use actix_web::dev::Server;
 use actix_web::middleware::NormalizePath;
 use actix_web::{web, App, HttpServer};
-use shaku::{Component, Interface};
-use sqlx::mysql::{MySqlConnectOptions, MySqlPoolOptions};
-use sqlx::MySqlPool;
 
+use crate::repositories::Database;
 use crate::routes::{health_check, signup};
-use crate::AppModule;
+use crate::{AppModule, MySqlDatabase};
 
 pub struct ApiServer {
     port: u16,
@@ -47,44 +44,5 @@ impl ApiServer {
 
     pub fn port(&self) -> u16 {
         self.port
-    }
-}
-
-pub trait Database: Interface {
-    fn pool(&self) -> &MySqlPool;
-}
-
-#[derive(Component)]
-#[shaku(interface = Database)]
-pub struct MySqlDatabase {
-    pool: MySqlPool,
-}
-
-impl MySqlDatabase {
-    pub fn new() -> Self {
-        let options = MySqlConnectOptions::from_str(
-            std::env::var("DATABASE_URL")
-                .expect("DATABASE_URL must be set")
-                .as_str(),
-        )
-        .unwrap();
-
-        let pool = MySqlPoolOptions::new()
-            .max_connections(5)
-            .connect_lazy_with(options);
-
-        Self { pool }
-    }
-}
-
-impl Default for MySqlDatabase {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-impl Database for MySqlDatabase {
-    fn pool(&self) -> &MySqlPool {
-        &self.pool
     }
 }
